making notes and TODOs
Not sure everything should go into runtime.

It might be nice if we could put some of the logic that's repeatable into the Scenario Model itself
so that it knows how to execute itself, using the Glue. Not sure. I looked at the Junit stuff and
my brain started to melt, so I figured I will start at regular CLI runtime first, and then get the Junit stuff
to behave.