Tell don't ask when printing summary at the end of the run
Let the SummaryPrinter tell the Runtime to print the summary instead of
asking for the SummaryCounter.

Put back the check for an emtpy list of backends in the Runtime ctor,
which was removed by mistake. Add test to check that a CucumberException
is thrown when the list of backends is emtpy.

Use Long for duration arguments in SummaryCounter, to conform to the
return type of Result.getDuration().