The JSONFormatter should record before hooks in the next scenario
For all but the before hooks executed before the first scenario, the
JSONFormatter records them for the latest started scenario, instead of
the scenario be to executed next. The before hook data is shifted up
one scenario, so generally the first scenario will record two hook
executions for each before hook, and the last scenario none.