Only include executed scenarios and steps in the JSON output
Do not include the Scenario Outline and Examples given as information
before the execution of the instantiated scenarios from the outline.
This unified the JSON output from Cucumber-JVM regardless of the feature
is executed by the command line runner, the JUnit runner of the TestNG
runner. The JUnit runner differs from the other by not sending the
Scenario Outline and Examples as information to the formatters before
executing the instantiated scenarios from the outline, so this commit
only affects the output from the command line runner and the TestNG
runner.