Propagate parse error into the TestNG test execution. Fixes #953.
The CucumberException thrown from feature file parse errors is silently
caught when using a TestNG DataProvider to run each feature as a
separate TestNG test. Let the DataProvider wrap that parse error in a
object so the error can be rethrown inside the TestNG test case, to make
TestNG aware of the parse error.