Map AssumptionViolatedException to Skipped status.
AssumptionViolatedExceptions are used in JUnit to signal after the test
has started that it is not meaningful given the current conditions.
Currently these exceptions are mapped to the Pending status, but it is
a better match to map them to the Skipped status, so that the exit code
is zero in the presence of them also in the strict mode.