Extract TableConverter interface
TableConverter is not part of the public api but is used by DataTable.
By extracting the interface we can hide the details of how a data table
converts its data.

This will also be usefull once we start moving away from xstream as we
can introduce an XStreamTableConverter and SomeOtherTableConverter.