[Core] Set scenario result as step finishes (#1430)
* Clean up pom

* Update scenario result when step finishes

The results of a step were set on the scenario after a test step
finished. This meant that the result of the step was not available in
its after hooks.

By adding the result of the step to the scenario in TestStep we ensure
that scenario always contains the most recent result. This removes the
need to pass the scenario results around.

Additionally:

 * Scenario has moved from cucumber.runtime.ScenarioImpl to
 cucumber.runner.Scenario and made package private. This prevents the
 implementation being used outside of the API. Users of
 ScenarioImpl.getError should use an event listener to create their
 reports.

 * TestStep.run no longer requires a language parameter. With Gherkin v5
 the language is not used anywhere.