[Core] Share object factory between backends
By sharing the object factory between different backends it becomes
possible to use the same test context in different languages. This is
very useful when mixing Kotlin and Java, or Java and Java 8.

This also requires that the backend no longer manages the object factory
life-cycle. To end a container and lookup have been extracted from the
object factory.

Closes #1117.