[Core] Add object-factory option to CLI and CucumberOptions (#1710)
Summary

In order to provide different test configurations it might be useful to use different object factories when working with
a DI framework. This is described in issue #1690.

Details

The annotations for JUnit and TestNG now allow to specify the class to instantiate as object factory. A CLI flag has
been added to allow the specification of an object factory on the command line (--object-factory). Some minor
refactorings along the way.

Motivation and Context

It is now possible to use different object factories in a Cucumber test suite. Features need to be separated (using
tags) and can be grouped as separate tests that are annotated with different CucumberOptions. For each specification of
CucumberOptions a different objectFactory can be specified. That way different feature tests can e.g. use differently
configured injectors. It is no longer necessary to specify an object factory in cucumber.properties. Object factories
are loaded via the ServiceLoader and need to be specified in the file
META-INF/services/io.cucumber.core.backend.ObjectFactory. Multiple implementations can be given here.