[Core] Pull tag-expression matching into core
Each hook definition implementation was required to parse tag expressions
and match these against pickle tags. This creates some duplication of
effort. Most of this duplication was remedied by reusing the
`PicklePredicate`. However the `PicklePredicate` is not part of the API
exposed by the core. Moving the parsing and and matching of pickle tags
into the core resolves all these issues.