[Core] Add dedicated DocStringTypeRegistry (#1745)
Cucumber uses the `DatatTableTypeRegistry` to convert doc strings. This is
rather confusing and ignores the type provided with the doc string. This
introduces a dedicated `DocStringTypeRegistry`.

Given some gherkin

```
Given a doc string
"""json
{hello:"world"}
"""
```

```
And a step definition:

```
@Given("a doc string")
public void a_doc_string(JsonNode json){

}

```

The using the `@DocStringType` annotation it is possible

```
public JsonNode json(String docString){
  return objectMapper.readTree(docString)
}
```

to convert the doc string automatically to `JsonNode`.