[Core] Remove detail parameter from getLocation
All glue definition classes implemented a `getLocation(boolean detail)`
method. This would provide a long form method description. However
in practice this method was always invoked with true. Removing the
parameter cleans up the API.

This also revealed that `Invoker` is only used for the `timeout` in
`cucumber-java8`. This means that class can be moved to `cucumber-java`
and with that `MethodFormat` can also be moved to `cucumber-java`.