[Core] Include name in duplicate feature detection (#1819)
Build tools will typically copy the source files to a `build` or
`target` directory duplicating all feature files in the project. When
the root of the project is used as the feature path cucumber will
discover both sets of feature files (https://github.com/cucumber/cucumber-jvm/issues/165).

By including the file name we can distinguish between accidental
duplication and intentional duplication.

We could make this scenario an error but given the complexities of https://github.com/cucumber/cucumber-jvm/issues/259
this is best left as a separate issue.