[JUnit] Map SKIPPED to TestAbortedException
The semantics of `TestSkippedException` require that the test is skipped
before the JUnit Platform executes the test. The semantics of
`TestAbortedException` require that the test has been started and is
then aborted. So because cucumber runs within the JUnit Platform and
will not discover that a test should be skipped until it is executed
this follows the `TestAbortedException` semantics despite the confusing
names.