[Core] Extract Gherkin compatibility layer (#1804)
Gherkin 6 introduced the `Rule` keyword and a new AST structure. This poses
several problems.

1. Cucumber-JVM is closely tied to the Pickle structure of Gherkin 5.
2. The HTML and JSON formatters use the Gherkin 5 parser.
3. The JSON formatter is the defacto output standard for third party tools.
4. There is no schema for the JSON formatters output.

To phase out the JSON formatter we'll need an alternative. This alternative
is the `message` formatter. This plugin will write the output of Cucumbers
execution to protobuf or ndjson file using the schema defined in
`cucumber-messages`.

Because `cucumber-messages` for Gherkin can only be generated by Gherkin 8
we need a way to run both Gherkin 5 and Gherkin 8 next to each other. By
extracting a compatibility layer we can use both Gherkin 5 and Gherkin 8.